# EcommerceAPI.postman_collection.json-Swagger20.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "EcommerceAPI",
    "contact": {}
  },
  "host": "localhost:8080",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/item/createItem": {
      "post": {
        "summary": "create item",
        "tags": [
          "item"
        ],
        "operationId": "createitem",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "itemName",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "barcode",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "importPrice",
            "in": "formData",
            "required": true,
            "type": "integer",
            "format": "int32",
            "description": ""
          },
          {
            "name": "price",
            "in": "formData",
            "required": true,
            "type": "integer",
            "format": "int32",
            "description": ""
          },
          {
            "name": "weight",
            "in": "formData",
            "required": true,
            "type": "integer",
            "format": "int32",
            "description": ""
          },
          {
            "name": "thumbnail",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "description",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "numberWare",
            "in": "formData",
            "required": true,
            "type": "integer",
            "format": "int32",
            "description": ""
          },
          {
            "name": "categoryId",
            "in": "formData",
            "required": true,
            "type": "integer",
            "format": "int32",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/item/getItem": {
      "get": {
        "summary": "get item",
        "tags": [
          "item"
        ],
        "operationId": "getitem",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/item/getItem/1": {
      "get": {
        "summary": "get item id",
        "tags": [
          "item"
        ],
        "operationId": "getitemid",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/item/updateItem/1": {
      "put": {
        "summary": "update item",
        "tags": [
          "item"
        ],
        "operationId": "updateitem",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "parameters": [
          {
            "name": "itemname",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "barcode",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "importPrice",
            "in": "formData",
            "required": true,
            "type": "integer",
            "format": "int32",
            "description": ""
          },
          {
            "name": "price",
            "in": "formData",
            "required": true,
            "type": "integer",
            "format": "int32",
            "description": ""
          },
          {
            "name": "weight",
            "in": "formData",
            "required": true,
            "type": "number",
            "format": "double",
            "description": ""
          },
          {
            "name": "thumbnail",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "description",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "numberWare",
            "in": "formData",
            "required": true,
            "type": "integer",
            "format": "int32",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
    "/orderDetail/createOD": {
      "post": {
        "summary": "create order detail",
        "tags": [
          "order detail"
        ],
        "operationId": "createorderdetail",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/createorderdetailrequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
    "/orderDetail/updateOD/2": {
      "put": {
        "summary": "update order detail",
        "tags": [
          "order detail"
        ],
        "operationId": "updateorderdetail",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/updateorderdetailrequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
    "/order/createOrder": {
      "post": {
        "summary": "create order",
        "tags": [
          "order"
        ],
        "operationId": "createorder",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/createorderrequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
    "/order/getOrder/1": {
      "get": {
        "summary": "get order id",
        "tags": [
          "order"
        ],
        "operationId": "getorderid",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "createorderdetailrequest": {
      "title": "createorderdetailrequest",
      "example": {
        "quantity": 1,
        "orderId": 1,
        "itemId": 2
      },
      "type": "object",
      "properties": {
        "quantity": {
          "type": "integer",
          "format": "int32"
        },
        "orderId": {
          "type": "integer",
          "format": "int32"
        },
        "itemId": {
          "type": "integer",
          "format": "int32"
        }
      },
      "required": [
        "quantity",
        "orderId",
        "itemId"
      ]
    },
    "updateorderdetailrequest": {
      "title": "updateorderdetailrequest",
      "example": {
        "price": 999,
        "quantity": 12
      },
      "type": "object",
      "properties": {
        "price": {
          "type": "integer",
          "format": "int32"
        },
        "quantity": {
          "type": "integer",
          "format": "int32"
        }
      },
      "required": [
        "price",
        "quantity"
      ]
    },
    "createorderrequest": {
      "title": "createorderrequest",
      "example": {
        "userId": 1,
        "orderDetailId": null,
        "codeVoucher": "ABC123"
      },
      "type": "object",
      "properties": {
        "userId": {
          "type": "integer",
          "format": "int32"
        },
        "orderDetailId": {
          "type": "string"
        },
        "codeVoucher": {
          "type": "string"
        }
      },
      "required": [
        "userId",
        "orderDetailId",
        "codeVoucher"
      ]
    },
...
  },
...
}